update run.sh so pip install only happens once

+++ run.sh
@@ -1,10 +1,19 @@
 #!/bin/bash
 cd `dirname $0`
 
-python3 -m pip install --user virtualenv
-python3 -m venv viam-env
-source viam-env/bin/activate
-pip3 install --upgrade -r requirements.txt
+if [ -f .installed ]
+  then
+    source viam-env/bin/activate
+  else
+    python3 -m pip install --user virtualenv
+    python3 -m venv viam-env
+    source viam-env/bin/activate
+    pip3 install --upgrade -r requirements.txt
+    if [ $? -eq 0 ]
+      then
+        touch .installed
+    fi
+fi
 
 # Be sure to use `exec` so that termination signals reach the python process,
 # or handle forwarding termination signals manually
